Dan #atomic #refac Rename to clarify role of Github object.
It's a file system.  And we're using an in memory file system in unit specs.
