+++ app/controllers/create_content_controller.rb
@@ -1,4 +1,4 @@
-require_relative '../github/github'
+require_relative '../github/github_file_system'
 require_relative '../utilities/log'
 require_relative '../model/content'
 require_relative '../exceptions/unprocessable_entity_error'
