+++ app/routes/content_service_routes.rb
@@ -4,7 +4,7 @@ require_relative '../../app/stores/file_data_store'
 require_relative '../web/request_error_handler'
 require_relative '../exceptions/page_not_found_error'
 require_relative '../web/wiring'
-require_relative '../../app/github/github'
+require_relative '../../app/github/github_file_system'
 require_relative '../../app/security/hmac_authorisation'
 require_relative '../../app/web/halt'
 require_relative '../web/wiring'
