+++ app/web/wiring.rb
@@ -5,7 +5,7 @@ require_relative '../utilities/file_system_wrapper'
 require_relative '../utilities/shell_executor'
 require_relative '../github/logging_github_client'
 require_relative '../github/github_client'
-require_relative '../github/github'
+require_relative '../github/github_file_system'
 require_relative '../lib/content_digest'
 require_relative '../controllers/create_content_controller'
 require_relative '../controllers/delete_content_controller'
@@ -42,8 +42,8 @@ module Hacienda
       ContentStore.new(:public, query(request.env['rack.request.query_hash']), content_handlers, content_file_store, log, local_content_repo)
     end
 
-    def github
-      Github.new(settings, logging_github_client)
+    def github_file_system
+      GithubFileSystem.new(settings, logging_github_client)
     end
 
     def logging_github_client
@@ -55,19 +55,19 @@ module Hacienda
     end
 
     def publish_content_controller
-      PublishContentController.new(github, content_digest, log)
+      PublishContentController.new(github_file_system, content_digest, log)
     end
 
     def create_content_controller
-      CreateContentController.new(github, content_digest)
+      CreateContentController.new(github_file_system, content_digest)
     end
 
     def delete_content_controller
-      DeleteContentController.new(github, log)
+      DeleteContentController.new(github_file_system, log)
     end
 
     def update_content_controller
-      UpdateContentController.new(github, content_digest, draft_content_store, log)
+      UpdateContentController.new(github_file_system, content_digest, draft_content_store, log)
     end
 
     def file_system_wrapper
