+++ spec/end_to_end/support/end_to_end_test_helper.rb
@@ -2,7 +2,7 @@ require 'rspec'
 
 require_relative 'faraday_client'
 require_relative '../../shared/hacienda_runner'
-require_relative '../../../app/github/github'
+require_relative '../../../app/github/github_file_system'
 require_relative '../../../app/github/github_client'
 require_relative '../../../app/exceptions/not_found_exception'
 require_relative '../../coverage_profiles'
