+++ spec/functional/content_deletion_spec.rb
@@ -3,7 +3,7 @@ require_relative 'support/test_content_manager'
 
 require_relative '../shared/metadata_builder'
 require_relative '../../spec/shared/navigation'
-require_relative '../../spec/functional/support/fake_github'
+require_relative '../../spec/functional/support/fake_github_file_system'
 require_relative '../shared/content_item'
 
 module Hacienda
@@ -12,13 +12,13 @@ module Hacienda
     describe 'deleting items' do
       include Navigation
 
-      let(:fake_github) { FakeGithub.new(TEST_REPO) }
+      let(:fake_github) { FakeGithubFileSystem.new(TEST_REPO) }
       let(:a_content_item) { ContentItem.new }
       let(:content) { Content.new }
 
       before :each do
         app.set :content_directory_path, TEST_REPO
-        allow_any_instance_of(app).to receive(:github).and_return(fake_github)
+        allow_any_instance_of(app).to receive(:github_file_system).and_return(fake_github)
       end
 
       it 'should delete all content for a specific item' do
