+++ spec/unit/web/wiring_spec.rb
@@ -59,7 +59,7 @@ module Hacienda
 
       it 'should return content service controller' do
         expect(app.create_content_controller).to be_a CreateContentController
-        expect(app.create_content_controller.instance_variable_get(:@file_system)).to be_a Github
+        expect(app.create_content_controller.instance_variable_get(:@file_system)).to be_a GithubFileSystem
       end
 
       it 'should return a content update controller using draft content store... cos public doesnt return version info
@@ -69,8 +69,8 @@ module Hacienda
       end
 
       it 'should log github_client calls' do
-        github = app.github
-        expect(github).to be_a Github
+        github = app.github_file_system
+        expect(github).to be_a GithubFileSystem
         expect(github.instance_variable_get(:@github_client)).to be_a LoggingGithubClient
       end
